# ApiDoc/DocumentationOpenApi-v3.0.1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ApiRest",
    "version": "1.0"
  },
  "paths": {
    "/api/report/one": {
      "get": {
        "tags": [
          "Get"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuelPercentageResponseRSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          }
        }
      }
    },
    "/api/report/two": {
      "get": {
        "tags": [
          "Get"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuelAverageResponseRSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          }
        }
      }
    },
    "/api/report/fourth": {
      "get": {
        "tags": [
          "Get"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FuelPercentageBodyRSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fuel": {
      "post": {
        "tags": [
          "Post"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FuelConsumptionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FuelConsumptionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FuelConsumptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanRSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          }
        }
      }
    },
    "/api/energy": {
      "post": {
        "tags": [
          "Post"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnergyConsumptionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EnergyConsumptionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EnergyConsumptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanRSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          }
        }
      }
    },
    "/api/petroleum": {
      "post": {
        "tags": [
          "Post"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetroleumConsumptionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PetroleumConsumptionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PetroleumConsumptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanRSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          }
        }
      }
    },
    "/api/travel": {
      "post": {
        "tags": [
          "Post"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TravelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanRSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BooleanRSuccess": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "EnergyConsumptionRequest": {
        "type": "object",
        "properties": {
          "units": {
            "type": "number",
            "format": "double"
          },
          "consumptionDate": {
            "type": "string",
            "format": "date-time"
          },
          "energyCatalogId": {
            "type": "integer",
            "format": "int32"
          },
          "issuanceCatalogId": {
            "type": "integer",
            "format": "int32"
          },
          "energyTypeCatalogId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "FuelAverageResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "units": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "FuelAverageResponseRSuccess": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/FuelAverageResponse"
          }
        },
        "additionalProperties": false
      },
      "FuelConsumptionRequest": {
        "type": "object",
        "properties": {
          "units": {
            "type": "number",
            "format": "double"
          },
          "consumptionDate": {
            "type": "string",
            "format": "date-time"
          },
          "fuelCatalogId": {
            "type": "integer",
            "format": "int32"
          },
          "issuanceCatalogId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "FuelPercentageBody": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "nullable": true
          },
          "percentage": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "FuelPercentageBodyRSuccess": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/FuelPercentageBody"
          }
        },
        "additionalProperties": false
      },
      "FuelPercentageResponse": {
        "type": "object",
        "properties": {
          "body": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FuelPercentageBody"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FuelPercentageResponseRSuccess": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/FuelPercentageResponse"
          }
        },
        "additionalProperties": false
      },
      "PetroleumConsumptionRequest": {
        "type": "object",
        "properties": {
          "units": {
            "type": "number",
            "format": "double"
          },
          "consumptionDate": {
            "type": "string",
            "format": "date-time"
          },
          "petroleumCatalogId": {
            "type": "integer",
            "format": "int32"
          },
          "issuanceCatalogId": {
            "type": "integer",
            "format": "int32"
          },
          "derivativeCatalogId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RError": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TravelRequest": {
        "type": "object",
        "properties": {
          "destination": {
            "type": "string",
            "nullable": true
          },
          "travelDate": {
            "type": "string",
            "format": "date-time"
          },
          "travelCatalogId": {
            "type": "integer",
            "format": "int32"
          },
          "issuanceCatalogId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}